backend/requirements.txt
flask~=2.1.3
twilio~=6.38.1
python-dotenv==0.13.0
pytest==5.4.1
Werkzeug==2.2.2